Refactor: Moved the Homepage's visit Students Page button to its own component
